2.2.4

+++ hetrixtools_agent.sh
@@ -24,7 +24,7 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 ScriptPath=$(dirname "${BASH_SOURCE[0]}")
 
 # Agent Version (do not change)
-Version="2.2.3"
+Version="2.2.4"
 
 # Load configuration file
 if [ -f "$ScriptPath"/hetrixtools.cfg ]
@@ -66,26 +66,41 @@ function base64prep() {
 	echo "$str"
 }
 
+# Debug function
+function debug() {
+	if [ "$DEBUG" -eq 1 ]
+	then
+		echo -e "$ScriptStartTime-$(date +%T]) $1" >> "$ScriptPath"/debug.log
+	fi
+}
+
+# Clear debug.log every day at midnight
+if [ -z "$(date +%H | sed 's/^0*//')" ] && [ -z "$(date +%M | sed 's/^0*//')" ] && [ -f "$ScriptPath"/debug.log ]
+then
+	rm -f "$ScriptPath"/debug.log
+	debug "Cleared debug.log"
+fi
+
 # Start timers
 START=$(date +%s)
 tTIMEDIFF=0
+
+# Get current minute
 M=$(date +%M | sed 's/^0*//')
 if [ -z "$M" ]
 then
+	# If minute is empty, set it to 0
 	M=0
-	# Clear the hetrixtools_cron.log every hour
-	rm -f "$ScriptPath"/hetrixtools_cron.log
-	# If debug.log exists, clear it as well
-	if [ -f "$ScriptPath"/debug.log ]
+
+	# Clear hetrixtools_cron.log every hour
+	if [ -f "$ScriptPath"/hetrixtools_cron.log ]
 	then
-		rm -f "$ScriptPath"/debug.log
+		rm -f "$ScriptPath"/hetrixtools_cron.log
+		debug "Cleared hetrixtools_cron.log"
 	fi
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Cleared logs" >> "$ScriptPath"/debug.log; fi
 fi
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Starting HetrixTools Agent v$Version" >> "$ScriptPath"/debug.log; fi
+debug "Starting HetrixTools Agent v$Version"
 
 # Kill any lingering agent processes
 HTProcesses=$(pgrep -f hetrixtools_agent.sh | wc -l)
@@ -93,24 +108,25 @@ if [ -z "$HTProcesses" ]
 then
 	HTProcesses=0
 fi
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Found $HTProcesses agent processes\n$(ps aux | grep 'hetrixtools_agent.sh')" >> "$ScriptPath"/debug.log; fi
+debug "Found $HTProcesses agent processes\n$(ps aux | grep 'hetrixtools_agent.sh')"
+
 if [ "$HTProcesses" -ge 50 ]
 then
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Killing $HTProcesses lingering agent processes" >> "$ScriptPath"/debug.log; fi
+	debug "Killing $HTProcesses lingering agent processes"
 	pgrep -f hetrixtools_agent.sh | xargs kill -9
 fi
-for PID in $(pgrep -f hetrixtools_agent.sh)
-do
-	PID_TIME=$(ps -p "$PID" -oetime= | tr '-' ':' | awk -F: '{total=0; m=1;} {for (i=0; i < NF; i++) {total += $(NF-i)*m; m *= i >= 2 ? 24 : 60 }} {print total}')
-	if [ -n "$PID_TIME" ] && [ "$PID_TIME" -ge 90 ]
-	then
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Killing PID $PID, running for $PID_TIME seconds" >> "$ScriptPath"/debug.log; fi
-		kill -9 "$PID"
-	fi
-done
+if [ "$HTProcesses" -ge 10 ]
+then
+	for PID in $(pgrep -f hetrixtools_agent.sh)
+	do
+		PID_TIME=$(ps -p "$PID" -oetime= | tr '-' ':' | awk -F: '{total=0; m=1;} {for (i=0; i < NF; i++) {total += $(NF-i)*m; m *= i >= 2 ? 24 : 60 }} {print total}')
+		if [ -n "$PID_TIME" ] && [ "$PID_TIME" -ge 90 ]
+		then
+			debug "Killing PID $PID, running for $PID_TIME seconds"
+			kill -9 "$PID"
+		fi
+	done
+fi
 
 # Network interfaces
 if [ -n "$NetworkInterfaces" ]
@@ -122,8 +138,7 @@ else
 	NetworkInterfacesArray=()
 	while IFS='' read -r line; do NetworkInterfacesArray+=("$line"); done < <(ip a | grep BROADCAST | grep 'state UP' | awk '{print $2}' | awk -F ":" '{print $1}' | awk -F "@" '{print $1}')
 fi
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Network Interfaces:" "${NetworkInterfacesArray[@]}" >> "$ScriptPath"/debug.log; fi
+debug "Network Interfaces: ${NetworkInterfacesArray[*]}"
 
 # Initial network usage
 T=$(cat /proc/net/dev)
@@ -137,8 +152,7 @@ for NIC in "${NetworkInterfacesArray[@]}"
 do
 	aRX[$NIC]=$(echo "$T" | grep -w "$NIC:" | awk '{print $2}')
 	aTX[$NIC]=$(echo "$T" | grep -w "$NIC:" | awk '{print $10}')
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Network Interface $NIC RX: ${aRX[$NIC]} TX: ${aTX[$NIC]}" >> "$ScriptPath"/debug.log; fi
+	debug "Network Interface $NIC RX: ${aRX[$NIC]} TX: ${aTX[$NIC]}"
 done
 
 # Port connections
@@ -150,8 +164,7 @@ then
 	for cPort in "${ConnectionPortsArray[@]}"
 	do
 		Connections[$cPort]=$(echo "$netstat" | grep -c ":$cPort$")
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Port $cPort Connections: ${Connections[$cPort]}" >> "$ScriptPath"/debug.log; fi
+		debug "Port $cPort Connections: ${Connections[$cPort]}"
 	done
 fi
 
@@ -167,8 +180,7 @@ then
 	for i in "${CheckServicesArray[@]}"
 	do
 		SRVCSR[$i]=$(( ${SRVCSR[$i]} + $(servicestatus "$i") ))
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Service $i status: ${SRVCSR[$i]}" >> "$ScriptPath"/debug.log; fi
+		debug "Service $i status: ${SRVCSR[$i]}"
 	done
 fi
 
@@ -177,8 +189,7 @@ declare -A vDISKs
 for i in $(timeout 3 df | awk '$1 ~ /\// {print}' | awk '{print $(NF)}')
 do
 	vDISKs[$i]=$(lsblk -l | grep -w "$i" | awk '{print $1}')
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Disk $i: ${vDISKs[$i]}" >> "$ScriptPath"/debug.log; fi
+	debug "Disk $i: ${vDISKs[$i]}"
 done
 declare -A IOPSRead
 declare -A IOPSWrite
@@ -187,24 +198,19 @@ for i in "${!vDISKs[@]}"
 do
 	IOPSRead[$i]=$(echo "$diskstats" | grep -w "${vDISKs[$i]}" | awk '{print $6}')
 	IOPSWrite[$i]=$(echo "$diskstats" | grep -w "${vDISKs[$i]}" | awk '{print $10}')
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Disk $i IOPS Read: ${IOPSRead[$i]} Write: ${IOPSWrite[$i]}" >> "$ScriptPath"/debug.log; fi
+	debug "Disk $i IOPS Read: ${IOPSRead[$i]} Write: ${IOPSWrite[$i]}"
 done
 
 # Calculate how many how many data sample loops
 RunTimes=$(echo | awk "{print 60 / $CollectEveryXSeconds}")
-
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Collecting data for $RunTimes loops" >> "$ScriptPath"/debug.log; fi
+debug "Collecting data for $RunTimes loops"
 
 # Collect data loop
 for X in $(seq "$RunTimes")
 do
 	# Get vmstat
 	VMSTAT=$(vmstat "$CollectEveryXSeconds" 2 | tail -1)
-
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) $VMSTAT" >> "$ScriptPath"/debug.log; fi
+	debug "$VMSTAT"
 	
 	# CPU usage
 	CPU=$(echo "$VMSTAT" | awk '{print 100 - $15}')
@@ -243,8 +249,7 @@ do
 	loadavg15=$(echo "$loadavg" | awk '{print $3}')
 	tloadavg15=$(echo | awk "{print $tloadavg15 + $loadavg15}")
 
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) CPU: $CPU IO wait: $CPUwa Steal time: $CPUst User time: $CPUus System time: $CPUsy" >> "$ScriptPath"/debug.log; fi
+	debug "CPU: $CPU IO wait: $CPUwa Steal time: $CPUst User time: $CPUus System time: $CPUsy Load: $loadavg1 $loadavg5 $loadavg15"
 	
 	# RAM usage
 	aRAM=$(echo "$VMSTAT" | awk '{print $4 + $5 + $6}')
@@ -274,8 +279,7 @@ do
 	RAMCache=$(echo | awk "{print $aRAMCache * 100 / $bRAM}")
 	tRAMCache=$(echo | awk "{print $tRAMCache + $RAMCache}")
 
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) RAM: $RAM Swap: $RAMSwap Buffers: $RAMBuff Cache: $RAMCache" >> "$ScriptPath"/debug.log; fi
+	debug "RAM: $RAM Swap: $RAMSwap Buffers: $RAMBuff Cache: $RAMCache"
 	
 	# Network usage
 	T=$(cat /proc/net/dev)
@@ -303,8 +307,7 @@ do
 		tTX[$NIC]=$(echo "${tTX[$NIC]}" | awk '{printf "%18.0f",$1}' | xargs)
 	done
 
-	# DEBUG
-	if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Network Traffic:" "${tRX[@]}" "${tTX[@]}" >> "$ScriptPath"/debug.log; fi
+	debug "Network Traffic: ${tRX[*]} ${tTX[*]}"
 	
 	# Port connections
 	if [ -n "$ConnectionPorts" ]
@@ -314,8 +317,7 @@ do
 		do
 			Connections[$cPort]=$(echo | awk "{print ${Connections[$cPort]} + $(echo "$netstat" | grep -c ":$cPort$")}")
 		done
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Port Connections:" "${Connections[@]}" >> "$ScriptPath"/debug.log; fi
+		debug "Port Connections: ${Connections[*]}"
 	fi
 
 	# Temperature
@@ -332,8 +334,7 @@ do
 				TempArrayCnt[$TempName]=$((TempArrayCnt[$TempName] + 1))
 				TempNameCnt=$((TempNameCnt + 1))
 		done
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Temperature thermal_zone:" "${TempArray[@]}" >> "$ScriptPath"/debug.log; fi
+		debug "Temperature thermal_zone: ${TempArray[*]}"
 	else
 		if command -v "sensors" > /dev/null 2>&1
 		then
@@ -357,8 +358,7 @@ do
 					fi
 				fi
 			done
-			# DEBUG
-			if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Temperature sensors:" "${TempArray[@]}" >> "$ScriptPath"/debug.log; fi
+			debug "Temperature sensors: ${TempArray[*]}"
 		else
 			if command -v "ipmitool" > /dev/null 2>&1
 			then
@@ -377,8 +377,7 @@ do
 						fi
 					fi
 				done
-				# DEBUG
-				if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Temperature ipmitool:" "${TempArray[@]}" >> "$ScriptPath"/debug.log; fi
+				debug "Temperature ipmitool: ${TempArray[*]}"
 			fi
 		fi
 	fi
@@ -391,8 +390,7 @@ do
 	fi
 	if [ "$MM" -ne "$M" ] 
 	then
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Minute changed, ending loop" >> "$ScriptPath"/debug.log; fi
+		debug "Minute changed, ending loop"
 		break
 	fi
 done
@@ -445,14 +443,13 @@ Hostname=$(uname -n | base64 | xargs | sed 's/ //g')
 # Server uptime
 Uptime=$(awk '{print $1}' < /proc/uptime | awk '{printf "%18.0f",$1}' | xargs)
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) User: $User OS: $OS Kernel: $Kernel Hostname: $Hostname Uptime: $Uptime" >> "$ScriptPath"/debug.log; fi
+debug "User: $User OS: $OS Kernel: $Kernel Hostname: $Hostname Uptime: $Uptime"
 
 # lscpu
 lscpu=$(lscpu)
 
 # CPU model
-CPUModel=$(grep -m1 'model name' /proc/cpuinfo | awk -F": " '{print $NF}' | xargs)
+CPUModel=$(grep -m1 -E 'model name|cpu model' /proc/cpuinfo | awk -F": " '{print $NF}' | xargs)
 if [ -z "$CPUModel" ]
 then
 	CPUModel=$(echo "$lscpu" | grep "^Model name:" | awk -F": " '{print $NF}' | xargs)
@@ -463,10 +460,10 @@ CPUModel=$(echo -ne "$CPUModel" | base64 | xargs | sed 's/ //g')
 CPUSockets=$(grep -i "physical id" /proc/cpuinfo | sort -u | wc -l)
 
 # CPU cores
-CPUCores=$(echo "$lscpu" | grep "^CPU(s):" | awk '{print $(NF)}')
+CPUCores=$(echo "$lscpu" | grep "^CPU(s):" | awk '{print $(NF)}' | xargs)
 
 # CPU threads
-CPUThreads=$(echo "$lscpu" | grep "^Thread(s) per core:" | awk '{print $(NF)}')
+CPUThreads=$(echo "$lscpu" | grep "^Thread(s) per core:" | awk '{print $(NF)}' | xargs)
 
 # CPU clock speed
 if [ -z "$tCPUSpeed" ] || [ "$tCPUSpeed" -eq 0 ]
@@ -496,8 +493,7 @@ loadavg1=$(echo | awk "{print $tloadavg1 / $X}")
 loadavg5=$(echo | awk "{print $tloadavg5 / $X}")
 loadavg15=$(echo | awk "{print $tloadavg15 / $X}")
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) CPU Model: $CPUModel Sockets: $CPUSockets Cores: $CPUCores Threads: $CPUThreads Speed: $CPUSpeed CPU: $CPU IO wait: $CPUwa Steal time: $CPUst User time: $CPUus System time: $CPUsy Load: $loadavg1 $loadavg5 $loadavg15" >> "$ScriptPath"/debug.log; fi
+debug "CPU Model: $CPUModel Sockets: $CPUSockets Cores: $CPUCores Threads: $CPUThreads Speed: $CPUSpeed CPU: $CPU IO wait: $CPUwa Steal time: $CPUst User time: $CPUus System time: $CPUsy Load: $loadavg1 $loadavg5 $loadavg15"
 
 # RAM size
 RAMSize=$(grep ^MemTotal: /proc/meminfo | awk '{print $2}')
@@ -522,8 +518,7 @@ RAMBuff=$(echo | awk "{print $tRAMBuff / $X}")
 # RAM cache usage
 RAMCache=$(echo | awk "{print $tRAMCache / $X}")
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) RAM Size: $RAMSize Usage: $RAM Swap Size: $RAMSwapSize Usage: $RAMSwap Buffers: $RAMBuff Cache: $RAMCache" >> "$ScriptPath"/debug.log; fi
+debug "RAM Size: $RAMSize Usage: $RAM Swap Size: $RAMSwapSize Usage: $RAMSwap Buffers: $RAMBuff Cache: $RAMCache"
 
 # Disks usage
 DISKs=$(echo -ne "$(timeout 3 df -TPB1 | sed 1d | grep -v -E 'tmpfs' | awk '{print $(NF)","$2","$3","$4","$5";"}')" | xargs | sed 's/ //g' | base64 | xargs | sed 's/ //g')
@@ -544,8 +539,7 @@ do
 done
 IOPS=$(echo -ne "$IOPS" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) lsblk: $(lsblk -lp | base64 | xargs | sed 's/ //g') Disks: $DISKs Inodes: $INODEs IOPS: $IOPS" >> "$ScriptPath"/debug.log; fi
+debug "lsblk: $(lsblk -lp | base64 | xargs | sed 's/ //g') Disks: $DISKs Inodes: $INODEs IOPS: $IOPS"
 
 # Total network usage and IP addresses
 RX=0
@@ -569,8 +563,7 @@ NICS=$(echo -ne "$NICS" | base64 | xargs | sed 's/ //g')
 IPv4=$(echo -ne "$IPv4" | base64 | xargs | sed 's/ //g')
 IPv6=$(echo -ne "$IPv6" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Network Interfaces: $NICS IPv4: $IPv4 IPv6: $IPv6" >> "$ScriptPath"/debug.log; fi
+debug "Network Interfaces: $NICS IPv4: $IPv4 IPv6: $IPv6"
 
 # Port connections
 CONN=""
@@ -585,8 +578,7 @@ then
 fi
 CONN=$(echo -ne "$CONN" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Port Connections: $CONN" >> "$ScriptPath"/debug.log; fi
+debug "Port Connections: $CONN"
 
 # Temperature
 TEMP=""
@@ -601,8 +593,7 @@ then
 fi
 TEMP=$(echo -ne "$TEMP" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Temperature: $TEMP" >> "$ScriptPath"/debug.log; fi
+debug "Temperature: $TEMP"
 
 # Check Services
 SRVCS=""
@@ -621,8 +612,7 @@ then
 fi
 SRVCS=$(echo -ne "$SRVCS" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) Services: $SRVCS" >> "$ScriptPath"/debug.log; fi
+debug "Services: $SRVCS"
 
 # Check Software RAID
 RAID=""
@@ -637,8 +627,7 @@ then
 		if grep -q "$ii " <<< "$mdstat"
 		then
 			mdadm=$(mdadm -D "$i")
-			# DEBUG
-			if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) mdadm -D $i:\n$mdadm" >> "$ScriptPath"/debug.log; fi
+			debug "mdadm -D $i:\n$mdadm"
 			if [ -n "$mdadm" ]
 			then
 				mnt=$(echo -ne "$dfPB1" | grep "$i " | awk '{print $(NF)}')
@@ -649,8 +638,7 @@ then
 	if [ -x "$(command -v zpool)" ]
 	then
 		zpoolsoverall=$(zpool status 2>/dev/null)
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) zpool status:\n$zpoolsoverall" >> "$ScriptPath"/debug.log; fi
+		debug "zpool status:\n$zpoolsoverall"
 		if ! grep -q "no pools available" <<< "$zpoolsoverall"
 		then
 			zpools=()
@@ -668,8 +656,7 @@ fi
 RAID=$(echo -ne "$RAID" | base64 | xargs | sed 's/ //g')
 ZP=$(echo -ne "$ZP" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) RAID: $RAID ZP: $ZP" >> "$ScriptPath"/debug.log; fi
+debug "RAID: $RAID ZP: $ZP"
 
 # Check Drive Health
 DH=""
@@ -680,8 +667,7 @@ then
 		for i in $(lsblk -lp | grep ' disk' | awk '{print $1}')
 		do
 			DHealth=$(smartctl -A "$i")
-			# DEBUG
-			if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) smartctl -A $i:\n$DHealth" >> "$ScriptPath"/debug.log; fi
+			debug "smartctl -A $i:\n$DHealth"
 			if grep -q 'Attribute' <<< "$DHealth"
 			then
 				DHealth=$(smartctl -H "$i")"\n$DHealth"
@@ -720,8 +706,7 @@ then
 	if [ -x "$(command -v nvme)" ] #Using nvme-cli (for NVMe)
 	then
 		NVMeList="$(nvme list)"
-		# DEBUG
-		if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) NVMe List:\n$NVMeList" >> "$ScriptPath"/debug.log; fi
+		debug "NVMe List:\n$NVMeList"
 		for i in $(lsblk -lp | grep ' disk' | awk '{print $1}')
 		do
 			DHealth=$(nvme smart-log "$i")
@@ -743,8 +728,7 @@ then
 fi
 DH=$(echo -ne "$DH" | base64 | xargs | sed 's/ //g')
 
-# DEBUG
-if [ "$DEBUG" -eq 1 ]; then echo -e "$ScriptStartTime-$(date +%T]) DH: $DH" >> "$ScriptPath"/debug.log; fi
+debug "DH: $DH"
 
 # Custom Variables
 CV=""
@@ -792,12 +776,12 @@ jsoncomp=$(echo -ne "$json" | gzip -cf | base64 -w 0 | sed 's/ //g' | sed 's/\//
 # Save data to file
 echo "j=$jsoncomp" > "$ScriptPath"/hetrixtools_agent.log
 
-# DEBUG
 if [ "$DEBUG" -eq 1 ]
 then
+	echo -e "$ScriptStartTime-$(date +%T]) JSON:\n$json" >> "$ScriptPath"/debug.log
 	# Post data
 	echo -e "$ScriptStartTime-$(date +%T]) Posting data" >> "$ScriptPath"/debug.log
-	wget -v --retry-connrefused --waitretry=1 -t 3 -T 15 -O- --post-file="$ScriptPath/hetrixtools_agent.log" $SecuredConnection https://sm.hetrixtools.net/v2/ &>> "$ScriptPath"/debug.log 
+	wget -v --debug --retry-connrefused --waitretry=1 -t 3 -T 15 -O- --post-file="$ScriptPath/hetrixtools_agent.log" $SecuredConnection https://sm.hetrixtools.net/v2/ &>> "$ScriptPath"/debug.log 
 	echo -e "$ScriptStartTime-$(date +%T]) Data posted" >> "$ScriptPath"/debug.log
 else
 	# Post data
